Tracking rates: float protocol (SXRr/SXRd, GXRr/GXRd, GXAS), align driver/LX200/webserver/app
- MainUnit/UniversalMainUnit: SXRr,SXRd parse float (strtod); GXRr,GXRd,GXRh return float (%.10g)
- GXAS packet: bytes 40,44 float32 for current track rate; 48,52 int32 for stored
- ASCOM driver: send float for DeclinationRate/RightAscensionRate; read float from GXAS; remove lastSetRightAscensionRate
- MoveAxis: send integer arcsec/s to avoid firmware rejection; SiteLatitude/SiteLongitude clearer error when not at home/park
- LX200Client: getTrackRateRA/Dec(double&), parse float reply
- TeenAstroMountStatus: GXAS parse float at 40,44; members and getters double; getTrackingRate uses 1.0 - m_trackRateRa
- TeenAstroWifi Helper: use getTrackingRateRa/Dec directly (no /10000)
- teenastro_app: trackRateRa/Dec double, GXAS getFloat32 at 40,44
- CommandReplyLength/lx200_reply_lengths: GXRr,GXRd,GXRh reply length 24

Made-with: Cursor

+++ libraries/TeenAstroLX200io/src/LX200Client.h
@@ -149,8 +149,8 @@ public:
   // -----------------------------------------------------------------------
   //  Tracking rates (typed)
   // -----------------------------------------------------------------------
-  LX200RETURN getTrackRateRA(long& value);             // :GXRr#
-  LX200RETURN getTrackRateDec(long& value);            // :GXRd#
+  LX200RETURN getTrackRateRA(double& value);            // :GXRr# (float reply: ASCOM RA rate)
+  LX200RETURN getTrackRateDec(double& value);           // :GXRd# (float reply: arcsec/s)
   LX200RETURN getStoredTrackRateRA(long& value);       // :GXRe#
   LX200RETURN getStoredTrackRateDec(long& value);      // :GXRf#
